pyproject.toml
[project]
name = "amoeba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "pygame-ce>=2.5.7",
]
